+++ changelog.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Account management UI
-
+- Filter by transaction Account
 ### Changes
 - Upgraded to NextJS 15
 - Upgraded to TailwindCSS v4